Keep track of how many values got read
By limiting the loop to the number of read values the min/max/avg values
will be corretly calculated. The problem was that this loop takes the
width (e.g. value 135) and loops over the given values (most of them
will be 0 in the first iterations). The effect was that the min and avg
values were 0 most of the time (until enough values were collected).

Fixes #50