Rely on pkg-config to find the proper ncurses.h
On Ubuntu, /usr/include/ncursesw/ncurses.h is a symlink to ../ncurses.h.
On Gentoo, however, /usr/include/ncurses.h and
/usr/include/ncursesw/ncurses.h are different files, and only the latter
defines the wide character functions: we need to give the compiler the
proper -I flag.

For the sake of portability, ask pgk-config for the correct compiler
flags (including -I). Default to the flags used on Gentoo, which also
work on Ubuntu.