ttyplot.c: Stop using broken-on-Linux pselect :(
With both aggressive stdin input and signals coming in,
`pselect` sometimes delays signal delivery for multiple
seconds, to the point where even holding Ctrl+C down
for seconds will not be successful in terminating
the process.  I have witnessed this symptom for a
duration of 20+ seconds in practice.

To reproduce:
> # ./ttyplot < /dev/zero
> [hold Ctrl+C or press repeatedly]
> [see it take multiple seconds for the process to terminate (if at all)]

Related:
https://stackoverflow.com/questions/62315082/pselect-on-linux-does-not-deliver-signals-if-events-are-pending