Catch signals on Linux with signalfd
On Linux, pselect() does not reliably catch signals when there is high
pressure on the watched file descriptors. This was mostly fixed in
commit e94cecd6a2a7873992c9a71131329bf2669a4557 by calling pselect()
twice. This solution, however, leaves a race window between the two
calls.

Fix the pselect() issue without races by using signalfd(). This makes
signal delivery reliable. The call is only issued on Linux, as both the
pselect() issue and signalfd() are Linux-specific.