Add support for negative values.
A couple of drive-by changes are notable:

- Data arrays are now initialized with NaNs, so we don't assume 0s. This
  used to be harmless while data was nonnegative, but if a function goes
  negative while initialized-with-0 data is still in the data buffer it
  becomes visibly wrong without this fix.
- The standard math library (-lm) is now used. This could be finessed
  with biased truncation, but this is clearer.

Data is still plotted vertically "up from negative infinity", which is
perhaps a little unusual. This is simpler than tracking where the axes
should be drawn, and still produces a legible result with axes that
don't jump around as the extents of the data change.