Store the version string in a single place
Instead of having the ttyplot version string in ttyplot.c (as three
separate variables) and recordings/expected.txt, store it only in
Makefile. This should make version bumping easier.

This string is given by `make` to the compiler as a -D option. The CI
workflow and get_back_in_sync.sh extract this string from the Makefile
using `grep`.