Fix maybe-uninitialized compiler warning
Some GCC versions say:

    ttyplot.c:359:16: error:
    ‘y2_start’ may be used uninitialized [-Werror=maybe-uninitialized]

and a similar warning (promoted to error) for y2_end. These are false
positives: the variables are initialized if they ever get used. It is,
however, easy to restructure the code in order to get the initialization
and usage of the variables under the same ‘if’ test, so let's silence
the warning this way.

Fixes: https://github.com/tenox7/ttyplot/actions/runs/17030161145/job/48271411839